# auth/Swagger/signup.swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Auth API",
    "version": "1.0.0"
  },
  "paths": {
    "/api/users/signup": {
      "post": {
        "summary": "Sign up a new user.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "The username of the new user."
                  },
                  "password": {
                    "type": "string",
                    "description": "The password of the new user."
                  },
                  "repeat_password": {
                    "type": "string",
                    "description": "The password of the new user, repeated for confirmation."
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "The email address of the new user."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Bad request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "The unique identifier of the user."
          },
          "username": {
            "type": "string",
            "description": "The username of the user."
          },
          "password": {
            "type": "string",
            "description": "The password of the user."
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "The email address of the user."
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "email has to have a valid email format. password has to be at least 8 characters and include capital,lowercase and special caracters. the repeat_password has to be exactly equal to password."
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "location": {
                  "type": "string",
                  "description": "The location of the validation error (e.g., \"body.username\")."
                },
                "message": {
                  "type": "string",
                  "description": "email has to have a valid email format. password has to be at least 8 characters and include capital,lowercase and special caracters. the repeat_password has to be exactly equal to password."
                }
              }
            }
          }
        }
      }
      
    }
  }
}